+++ NEWS.md
@@ -1,5 +1,7 @@
 # treesitter (development version)
 
+* Removed usage of non-API `ATTRIB()` and `SET_ATTRIB()` (#41).
+
 # treesitter 0.3.0
 
 * New `node_field_name_for_named_child()` (#35).